@@ -92,26 +93,22 @@
     }
   ],
   "devDependencies": {
-    "@types/glob": "^7.1.1",
-    "@types/mocha": "^7.0.2",
-    "@types/node": "^12.11.7",
-    "@typescript-eslint/eslint-plugin": "^2.18.0",
-    "@typescript-eslint/parser": "^2.34.0",
-    "eslint": "^6.8.0",
-    "glob": "^7.1.6",
-    "mocha": "^7.1.2",
-    "nyc": "^15.0.0",
-    "typescript": "^3.7.5",
-    "vscode-test": "^1.3.0"
+    "@types/mocha": "^10.0.6",
+    "@types/node": "^20.10.4",
+    "@typescript-eslint/eslint-plugin": "^6.14.0",
+    "@typescript-eslint/parser": "^6.14.0",
+    "eslint": "^8.56.0",
+    "mocha": "^10.2.0",
+    "npm-run-all": "^4.1.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   },
-  "dependencies": {},
   "activationEvents": [
     "onLanguage:groovy",
     "onLanguage:jenkinsfile",
     "workspaceContains:**/Jenkins*",
     "workspaceContains:**/*.nf",
-    "workspaceContains:**/*.gradle",
-    "onCommand:groovyLint.lint"
+    "workspaceContains:**/*.gradle"
   ],
   "main": "./client/out/extension",
   "contributes": {
