@@ -1,8 +1,8 @@
 import * as path from 'path';
-import { runTests } from 'vscode-test';
+import { runTests } from '@vscode/test-electron';
 const NYC = require('nyc');
 
-const codeCoverage = process.argv && process.argv.includes('--codecoverage');
+const codeCoverage = process?.argv.includes('--codecoverage');
 
 async function main() {
 	try {
@@ -14,43 +14,48 @@ async function main() {
 		// Passed to --extensionTestsPath
 		const extensionTestsPath = path.resolve(__dirname, './suite/index');
 
-		// create an nyc instance, config here is the same as in a package.json
-		const nycOptions = {
-			cwd: path.join(__dirname, '..', '..', '..'), // in debugging sessions, the cwd seems to be unset
-			include: [
-				"**/*.ts",
-				"**/*.js",
-			],
-			exclude: [
-				"coverage/**",
-				"node_modules/**",
-				"client/node_modules/**",
-				"server/node_modules/**",
-				"**/*.d.ts",
-				"**/*.test.ts"],
-			sourceMap: true,
-			extension: [
-				".ts"
-			],
-			all: true,
-			reporter: ['html'],
-			instrument: true,
-			hookRequire: true,
-			hookRunInContext: true,
-			hookRunInThisContext: true,
-		};
-		const nyc = new NYC(nycOptions);
+		let nyc;
 		if (codeCoverage) {
+			// create an nyc instance, config here is the same as in a package.json
+			const nycOptions = {
+				cwd: path.join(__dirname, '..', '..', '..'), // in debugging sessions, the cwd seems to be unset
+				include: [
+					"**/*.ts",
+					"**/*.js",
+				],
+				exclude: [
+					"coverage/**",
+					"node_modules/**",
+					"client/node_modules/**",
+					"server/node_modules/**",
+					"**/*.d.ts",
+					"**/*.test.ts"],
+				sourceMap: true,
+				extension: [
+					".ts"
+				],
+				all: true,
+				reporter: ['html'],
+				instrument: true,
+				hookRequire: true,
+				hookRunInContext: true,
+				hookRunInThisContext: true,
+			};
+			nyc = new NYC(nycOptions);
 			console.log('nyc options:\n' + JSON.stringify(nycOptions, null, 2));
 			nyc.createTempDirectory(); // create nyc' temp directory
 			nyc.wrap(); // hook into require() calls, etc.
 		}
+
 		// Download VS Code, unzip it and run the integration test
 		await runTests({
-			extensionPath: extensionDevelopmentPath,
-			testRunnerPath: extensionTestsPath,
-			additionalLaunchArgs: ['--disable-extensions'],
-			testRunnerEnv: (process.env["DEBUG"]) ? { "DEBUG": "vscode-groovy-lint,npm-groovy-lint" } : {}
+			extensionDevelopmentPath: extensionDevelopmentPath,
+			extensionTestsPath: extensionTestsPath,
+			extensionTestsEnv: {
+				DEBUG: 'vscode-groovy-lint,npm-groovy-lint',
+			},
+			launchArgs: ['--disable-extensions'],
+			reuseMachineInstall: true
 		});
 
 		if (codeCoverage) {
