@@ -5,7 +5,7 @@ import { NotificationType } from 'vscode-languageserver';
 // Status notifications
 export interface StatusParams {
 	id: number,
-	state: string, // lint.start, lint.start.format, lint.start.fix ,lint.end, lint.end.format, lint.end.fix, lint.error, lint.cancel
+	state: string, // lint.start, lint.start.format, lint.start.fix, lint.end, lint.end.format, lint.end.fix, lint.error, lint.cancel
 	documents: [
 		{
 			documentUri: string,
@@ -15,7 +15,7 @@ export interface StatusParams {
 	lastLintTimeMs?: number
 }
 export namespace StatusNotification {
-	export const type = new NotificationType<StatusParams, void>('groovylintlsp/status');
+	export const type = new NotificationType<StatusParams>('groovylintlsp/status');
 }
 
 // Active Document notifications to language server
@@ -23,7 +23,7 @@ export interface ActiveDocumentNotificationParams {
 	uri: string
 }
 export namespace ActiveDocumentNotification {
-	export const type = new NotificationType<ActiveDocumentNotificationParams, void>('groovylintlsp/activedocument');
+	export const type = new NotificationType<ActiveDocumentNotificationParams>('groovylintlsp/activedocument');
 }
 
 // Open textDocument or Url notification
@@ -34,7 +34,7 @@ export interface OpenNotificationParams {
 	preview?: boolean
 }
 export namespace OpenNotification {
-	export const type = new NotificationType<OpenNotificationParams, void>("groovylintlsp/open");
+	export const type = new NotificationType<OpenNotificationParams>("groovylintlsp/open");
 }
 
 // Usable settings
